commit
collider modifi + terrain

+++ Physics/PhysicsCommon.h
@@ -224,7 +224,7 @@ struct HeightFieldColliderInfo
 {
 	ColliderInfo colliderInfo;	//콜라이더 정보
 	//unsigned int heightFieldHash = 0;	//HeightField 해쉬값
-	int* heightMep = nullptr;	//모델 높이 맵 정보
+	float* heightMep = nullptr;	//모델 높이 맵 정보
 	unsigned int numCols = 0;	//높이 맵 가로 겟수
 	unsigned int numRows = 0;	//높이 맵 세로 겟수
 	float rowScale = 1.0f;	//높이 맵 세로 유닛 길이
